Refactor the metadata browser to use tabs. Adding an additional unit test.
