+++ iris-visualizer/src/main/java/cfa/vo/iris/visualizer/stil/IrisStarJTable.java
@@ -1,0 +1,31 @@
+/**
+ * Copyright (C) 2016 Smithsonian Astrophysical Observatory
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cfa.vo.iris.visualizer.stil;
+
+import uk.ac.starlink.table.gui.StarJTable;
+
+/**
+ * Simple bean wrapper for the StarJTable class.
+ *
+ */
+public class IrisStarJTable extends StarJTable {
+
+    public IrisStarJTable() {
+        super(false);
+    }
+
+}
