--- iris-visualizer/src/test/java/cfa/vo/iris/visualizer/metadata/MetadataBrowserViewTest.java
@@ -1,232 +1,0 @@
-/**
- * Copyright (C) 2016 Smithsonian Astrophysical Observatory
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package cfa.vo.iris.visualizer.metadata;
-
-import org.junit.Before;
-import org.junit.Test;
-import org.uispec4j.ListBox;
-import org.uispec4j.Panel;
-import org.uispec4j.Table;
-import org.uispec4j.Window;
-
-import cfa.vo.iris.IrisComponent;
-import cfa.vo.iris.sed.ExtSed;
-import cfa.vo.iris.sed.SedlibSedManager;
-import cfa.vo.iris.test.unit.AbstractComponentGUITest;
-import cfa.vo.iris.visualizer.VisualizerComponent;
-import cfa.vo.sedlib.Segment;
-import static org.junit.Assert.*;
-import static cfa.vo.iris.test.unit.TestUtils.*;
-
-import javax.swing.border.TitledBorder;
-
-public class MetadataBrowserViewTest extends AbstractComponentGUITest {
-
-    private VisualizerComponent comp = new VisualizerComponent();
-    private String plWindowName = comp.getName();
-    private String mbWindowName = MetadataBrowserView.MB_WINDOW_TITLE;
-
-    private MetadataBrowserView mbView;
-    private SedlibSedManager sedManager;
-    private Window mbWindow;
-
-    private Panel segmentPanel;
-    private ListBox segmentTable;
-    private Table metadataTable;
-
-    @Override
-    protected IrisComponent getComponent() {
-        return comp;
-    }
-
-    @Before
-    public void setupMbTest() throws Exception {
-        sedManager = (SedlibSedManager) app.getWorkspace().getSedManager();
-        
-        window.getMenuBar().getMenu("Tools").getSubMenu(plWindowName).getSubMenu(plWindowName).click();
-        
-        org.uispec4j.Button mbButton = desktop.getWindow(plWindowName).getButton("Metadata");
-        mbButton.click();
-        
-        desktop.containsWindow(mbWindowName).check();
-        mbWindow = desktop.getWindow(mbWindowName);
-        
-        mbView = comp.getDefaultPlotterView().getMetadataBrowserView();
-        mbView.reset();
-        
-        final Object[] tables = new Object[2];
-        invokeWithRetry(10, 100, new Runnable() {
-            @Override
-            public void run() {
-                tables[1] = mbWindow.getPanel("content").getPanel("segmentDataScrollPane").getTable("metadataTable");
-                tables[0] = mbWindow.getPanel("content").getPanel("segmentListScrollPane");
-            }
-        });
-        segmentPanel = (Panel) tables[0];
-        segmentTable = segmentPanel.getListBox();
-        metadataTable = (Table) tables[1];
-    }
-
-    @Test
-    public void testMetadataBrowser() throws Exception {
-        
-        // Load an sed into the workspace and ensure it shows up in the MB
-        final ExtSed sed = sedManager.newSed("test1");
-        sedManager.select(sed);
-        mbView.reset();
-        
-        // Verify the window title matches the selected SED id
-        TitledBorder sedTitle = (TitledBorder) mbView.segmentListScrollPane.getBorder();
-        assertEquals(sed.getId(), sedTitle.getTitle());
-
-        // Nothing should be selected
-        assertEquals(0, metadataTable.getRowCount());
-        assertEquals(0, mbView.selectedTables.getModel().getSize());
-        assertEquals(0, segmentTable.getSize());
-
-        // Add a segment to the selected sed
-        final Segment seg1 = createSampleSegment();
-        sed.addSegment(seg1);
-        
-        // 1 segment should have been added to table
-        invokeWithRetry(10, 100, new Runnable() {
-            @Override
-            public void run() {
-                assertEquals(1, mbView.selectedTables.getModel().getSize());
-            }
-        });
-        
-        
-        assertEquals(3, metadataTable.getRowCount());
-        assertEquals(1, Double.parseDouble((String) metadataTable.getContentAt(0, 1)), .1);
-
-        // Add another segment
-        double x[] = new double[] { 100, 200 };
-        double y[] = new double[] { 300, 400 };
-        final Segment seg2 = createSampleSegment(x, y);
-        sed.addSegment(seg2);
-        
-        invokeWithRetry(10, 100, new Runnable() {
-            @Override
-            public void run() {
-                // Verify there are two tables
-                assertEquals(2, mbView.selectedTables.getModel().getSize());
-            }
-        });
-        
-        // First segment should still be selected 
-        assertEquals(3, metadataTable.getRowCount());
-        assertEquals(1.0, mbView.selectedStarTable.getCell(0, 0));
-        assertEquals(1, Double.parseDouble((String) metadataTable.getContentAt(0, 1)), .1);
-        
-        // Set selected segment to seg2 by clicking on segment list
-        invokeWithRetry(10, 100, new Runnable() {
-            @Override
-            public void run() {
-                segmentTable.clearSelection();
-                segmentTable.click(1);
-            }
-        });
-        
-        // Second segment should be selected, verify table and value
-        assertEquals(100.0, mbView.selectedStarTable.getCell(0, 0));
-        assertEquals(2, metadataTable.getRowCount());
-        assertEquals(100, Double.parseDouble((String) metadataTable.getContentAt(0, 1)), .1);
-        
-        // Add a new SED
-        final ExtSed sed2 = sedManager.newSed("test2");
-        sedManager.select(sed2);
-        
-        invokeWithRetry(10, 100, new Runnable() {
-            @Override
-            public void run() {
-                // Verify the title has changed
-                TitledBorder sedTitle = (TitledBorder) mbView.segmentListScrollPane.getBorder();
-                assertEquals(sed2.getId(), sedTitle.getTitle());
-            }
-        });
-
-        // Verify the metadata and segment tables have cleared
-        assertEquals(0, metadataTable.getRowCount());
-        assertEquals(0, segmentTable.getSize());
-    }
-    
-    @Test
-    public void testMetadataBrowserEventListener() throws Exception {
-
-        // Load an sed into the workspace and ensure it shows up in the MB
-        final ExtSed sed = sedManager.newSed("testMetadataBrowserEventListener");
-        sedManager.select(sed);
-
-        // Nothing should be selected
-        assertEquals(0, metadataTable.getRowCount());
-        assertEquals(0, mbView.selectedTables.getModel().getSize());
-        assertEquals(0, segmentTable.getSize());
-        
-        invokeWithRetry(10, 100, new Runnable() {
-            @Override
-            public void run() {
-                // Verify title is correct
-                TitledBorder sedTitle = (TitledBorder) mbView.segmentListScrollPane.getBorder();
-                assertEquals(sed.getId(), sedTitle.getTitle());
-            }
-        });
-        
-        // Add a segment
-        final Segment seg1 = createSampleSegment();
-        sed.addSegment(seg1);
-        
-        // 1 segment should have been added to table
-        invokeWithRetry(10, 100, new Runnable() {
-            @Override
-            public void run() {
-                // Verify values in table are correct
-                assertEquals(1, mbView.selectedTables.getModel().getSize());
-                assertEquals(3, metadataTable.getRowCount());
-                assertEquals(1, Double.parseDouble((String) metadataTable.getContentAt(0, 1)), .1);
-            }
-        });
-        
-        // Add another segment
-        final Segment seg2 = createSampleSegment(new double[] {100}, new double[] {200});
-        sed.addSegment(seg2);
-        
-        // 1 segment should have been added to table
-        invokeWithRetry(10, 100, new Runnable() {
-            @Override
-            public void run() {
-                // Verify values in table are correct
-                assertEquals(2, mbView.selectedTables.getModel().getSize());
-                assertEquals(3, metadataTable.getRowCount());
-                assertEquals(1, Double.parseDouble((String) metadataTable.getContentAt(0, 1)), .1);
-            }
-        });
-        
-        // Remove the segment
-        sed.remove(seg1);
-        
-        // Segment should be removed from table
-        invokeWithRetry(10, 100, new Runnable() {
-            @Override
-            public void run() {
-                // Verify values in table are correct
-                assertEquals(1, mbView.selectedTables.getModel().getSize());
-                assertEquals(1, metadataTable.getRowCount());
-            }
-        });
-    }
-}
